# docs/swagger.yaml
---
openapi: 3.0.0
info:
  title: Тестовое API для CRUD операций Xsolla backend School
  contact:
    url: https://github.com/Reywaltz
    email: semenov.g.z@yandex.ru
  version: "1.0"
servers:
- url: http://localhost:8000/api/v1
paths:
  /items:
    get:
      tags:
      - Товары
      summary: Получение всех товаров
      description: Метод получения всех товаров с возможностью фильтрации
      operationId: GetItemsfromservice
      parameters:
      - name: type
        in: query
        description: Тип товара
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: min
        in: query
        description: Минимальная стоймость товара
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
      - name: max
        in: query
        description: Максимальная стоймость товара
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
      - name: limit
        in: query
        description: Количество получаемых данных
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
      - name: offset
        in: query
        description: Количество элементов, которые нужно пропустить перед началом сбора набора результатов
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                example:
                - sku: SON-SUB
                  name: Sonmsq
                  type: Sub
                  cost: "9999"
                - sku: POR-G-12
                  name: Portal
                  type: Game
                  cost: "12.23"
                - sku: SOS-SUB
                  name: Sosnmsq
                  type: Sub
                  cost: "9999.12"
                - sku: JUS-SUB
                  name: Justice
                  type: Sub
                  cost: "33221"
                - sku: JUS-GAM
                  name: Justice
                  type: Gam
                  cost: "332.2"
                - sku: JUX-GAM
                  name: Juxstice
                  type: Gam
                  cost: "332.2"
                - sku: JUX-SUB
                  name: Juxstice
                  type: Sub
                  cost: "3322"
                - sku: DOT-SUB
                  name: Juxstice
                  type: zxc
                  cost: "3322"
                - sku: COA-GAM
                  name: Coaol Videogame
                  type: Game
                  cost: "59.99"
                items:
                  $ref: '#/components/schemas/Getallitem'
              example:
              - sku: SON-SUB
                name: Sonmsq
                type: Sub
                cost: "9999"
              - sku: POR-G-12
                name: Portal
                type: Game
                cost: "12.23"
              - sku: SOS-SUB
                name: Sosnmsq
                type: Sub
                cost: "9999.12"
              - sku: JUS-SUB
                name: Justice
                type: Sub
                cost: "33221"
              - sku: JUS-GAM
                name: Justice
                type: Gam
                cost: "332.2"
              - sku: JUX-GAM
                name: Juxstice
                type: Gam
                cost: "332.2"
              - sku: JUX-SUB
                name: Juxstice
                type: Sub
                cost: "3322"
              - sku: DOT-SUB
                name: Juxstice
                type: zxc
                cost: "3322"
              - sku: COA-GAM
                name: Coaol Videogame
                type: Game
                cost: "59.99"
      deprecated: false
    post:
      tags:
      - Товары
      summary: Добавить товар
      description: Метод добавления нового товара
      operationId: Additem
      parameters: []
      requestBody:
        description: Данные нового товара
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdditemRequest'
            example:
              name: Name
              type: Item type
              cost: "0.00"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Successfullinsertion'
              example:
                sku: PRI-GAM
        "400":
          description: Входные данные неверного формата
        "409":
          description: Товар с данным SKU уже существует
      deprecated: false
  /items/{SKU}:
    get:
      tags:
      - Товары
      summary: Получение товара по индентификатору
      operationId: GetitembySKU
      parameters:
      - name: SKU
        in: path
        description: SKU товара
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "404":
          description: Товара с данным SKU не существует
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Successfulresponse'
              example:
                sku: SON-SUB
                name: Sonmsq
                type: Sub
                cost: "9999"
      deprecated: false
    put:
      tags:
      - Товары
      summary: Обновление товара
      description: Обновление данных товара по его идентификатору
      operationId: UpdateItem
      parameters:
      - name: SKU
        in: path
        description: SKU товара
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: Данные нового товара
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdditemRequest'
            example:
              name: New name
              type: New type
              cost: "0.00"
        required: true
      responses:
        "204":
          description: Товар обновлён
        "400":
          description: Неверные входные данные
    delete:
      tags:
      - Товары
      summary: Удаление товара
      description: Удаление товара по SKU
      operationId: DeleteItem
      parameters:
      - name: SKU
        in: path
        description: SKU товара
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "404":
          description: Товара не существует
        "204":
          description: Товар удалён
      deprecated: false
components:
  schemas:
    Getallitem:
      title: Getallitem
      required:
      - cost
      - name
      - sku
      - type
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        type:
          type: string
        cost:
          type: string
      example:
        sku: SON-SUB
        name: Sonmsq
        type: Sub
        cost: "9999"
    Successfulresponse:
      title: Successfulresponse
      required:
      - cost
      - name
      - sku
      - type
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
        type:
          type: string
        cost:
          type: string
      example:
        sku: SON-SUB
        name: Sonmsq
        type: Sub
        cost: "9999"
    AdditemRequest:
      title: AdditemRequest
      required:
      - cost
      - name
      - type
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        cost:
          type: string
      example:
        name: Name
        type: Item type
        cost: "0.00"
    Successfullinsertion:
      title: Successfullinsertion
      required:
      - sku
      type: object
      properties:
        sku:
          type: string
      example:
        sku: PRI-GAM